requirements.txt
# Core scientific stack
numpy>=2.2.3,<3.0
pandas>=2.2.3,<3.0
scipy>=1.15.2,<2.0
scikit-learn>=1.6.1,<2.0

# Bioinformatics + file I/O
biopython>=1.85,<2.0
h5py>=3.13.0,<4.0

# Deep Learning
torch>=2.6.0,<3.0
transformers>=4.49.0,<5.0
sentencepiece>=0.2.0

# Visualization
matplotlib>=3.10.1,<4.0
seaborn>=0.13.2,<1.0

# Utilities
tqdm>=4.67.1,<5.0
requests>=2.32.3,<3.0
pillow>=11.2.1,<12.0
regex>=2024.11.6

# Optional: Cloud support
boto3>=1.37.7
